make interactivemedical.org/imic2014 go to top level for now
We want imic2014 to be the cannonical site for this year's event
but don't have other content at the toplevel for now.  So we make
both URLs go to the same place and we can update the content
later by moving it to the subdirectory and putting new content
at the top.